refactor: 타입 파일명 오타 수정 및 타입 시스템 재구성
- patfair-types.ts → petfair-types.ts 파일명 오타 수정
- map-types.ts 분리로 지도 관련 타입 독립화
- PetFairSummaryData 타입 추가
- 관련 import 경로 모두 업데이트
